Inline ai-skills' manifest parser into ai-autopilot, remove the package
ai-autopilot's only real use of @gemstack/ai-skills was
parseSkillManifest (the SKILL.md frontmatter parser), for domain
presets and prompt bundles. Everything else in the package —
loadSkill/loadSkills, SkillRegistry, composeInstructions/Tools/
Middleware, SkillfulAgent (~470 of its ~540 lines) — had zero callers
anywhere in the repo.

Inlined the parser into packages/ai-autopilot/src/util/skill-manifest.ts
(with its tests) and dropped the workspace dependency, so the whole
package can go. Continues the token-burn cleanup from #1522.
