+++ packages/ai-autopilot/package.json
@@ -49,7 +49,7 @@
   },
   "dependencies": {
     "@gemstack/ai-sdk": "workspace:^",
-    "@gemstack/ai-skills": "workspace:^",
+    "yaml": "^2.5.0",
     "zod": "^4.0.0"
   },
   "peerDependencies": {
