+++ packages/ai-autopilot/src/preset/load.ts
@@ -1,8 +1,8 @@
 import { access, readFile, readdir } from 'node:fs/promises'
 import { fileURLToPath } from 'node:url'
 import { join } from 'node:path'
-import { parseSkillManifest } from '@gemstack/ai-skills'
-import type { SkillManifest } from '@gemstack/ai-skills'
+import { parseSkillManifest } from '../util/skill-manifest.js'
+import type { SkillManifest } from '../util/skill-manifest.js'
 import { defineLoop } from '../loop/define.js'
 import { parsePrompt } from '../prompts/parse.js'
 import type { Loop } from '../loop/types.js'
