--- packages/ai-skills/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@gemstack/ai-skills",
-  "version": "0.2.3",
-  "description": "Portable capability bundles for @gemstack/ai-sdk agents: load SKILL.md skills (instructions + tools + resources) and compose them onto an Agent.",
-  "keywords": [
-    "ai",
-    "agent",
-    "agents",
-    "skills",
-    "agent-skills",
-    "capabilities",
-    "tools",
-    "skill-md",
-    "gemstack"
-  ],
-  "license": "MIT",
-  "homepage": "https://github.com/gemstack-land/the-framework/tree/main/packages/ai-skills#readme",
-  "bugs": {
-    "url": "https://github.com/gemstack-land/the-framework/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gemstack-land/the-framework",
-    "directory": "packages/ai-skills"
-  },
-  "type": "module",
-  "engines": {
-    "node": ">=22.12.0"
-  },
-  "files": [
-    "dist"
-  ],
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "types": "./dist/index.d.ts"
-    }
-  },
-  "scripts": {
-    "build": "tsc -p tsconfig.build.json",
-    "dev": "tsc -p tsconfig.build.json --watch",
-    "typecheck": "tsc --noEmit",
-    "test": "tsc -p tsconfig.test.json && cd dist-test && node --test",
-    "clean": "rm -rf dist dist-test"
-  },
-  "dependencies": {
-    "@gemstack/ai-sdk": "workspace:^",
-    "yaml": "^2.5.0",
-    "zod": "^4.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.0.0",
-    "typescript": "^5.4.0"
-  },
-  "author": "Suleiman Shahbari"
-}
