--- packages/ai-skills/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": { "noEmit": true, "rootDir": "src" },
-  "include": ["src"]
-}
