--- packages/ai-skills/tsconfig.test.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": { "outDir": "dist-test", "rootDir": "src" },
-  "include": ["src"]
-}
